embedding-host/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pillow==11.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.90
